simulator/requirements.txt
PyYAML>=6.0
matplotlib>=3.0
reportlab>=4.0
